Admin-dashboard

+++ app/controllers/admin/users_controller.rb
@@ -1,5 +1,7 @@
 module Admin
   class UsersController < Admin::ApplicationController
+    before_action :require_admin
+
     expose(:users)
     expose(:user, attributes: :user_params)
 
@@ -14,23 +16,28 @@ module Admin
 
     def create
       user.save
-      respond_with(:admin, user)
+      respond_with :admin, user, location: -> { edit_admin_user_path(user) }
     end
 
     def update
       user.save
-      respond_with(:admin, user)
+      sign_in(user, bypass: true)
+      respond_with :admin, user, location: -> { edit_admin_user_path(user) }
     end
 
     def destroy
-      user.destroy
-      respond_with(:admin, user)
+      user.destroy unless current_admin_user == user
+      respond_with :admin, user, location: -> { admin_users_path }
     end
 
     private
 
     def user_params
-      params.require(:user).permit(:full_name, :email, :password)
+      params.require(:user).permit(:full_name, :email, :password, :password_confirmation)
+    end
+
+    def require_admin
+      redirect_to admin_root_path unless current_admin_user.admin?
     end
   end
 end
